backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.0
spacy>=3.7.0
chromadb>=0.4.24
openai>=1.14.0
tiktoken>=0.6.0
requests>=2.31.0
